rpg-11: Additive baseline-60 attribute standard

+++ src/main/java/com/keynor/rpg/application/dto/BloodSystemInput.java
@@ -2,7 +2,7 @@ package com.keynor.rpg.application.dto;
 
 import com.keynor.rpg.domain.model.BloodSystem;
 
-public record BloodSystemInput(double oxygenCarryingCapacity) {
+public record BloodSystemInput(int oxygenCarryingCapacity) {
 
     public BloodSystem toDomain() {
         return new BloodSystem(oxygenCarryingCapacity);
