+++ src/main/java/com/keynor/rpg/application/dto/BloodSystemResponse.java
@@ -2,7 +2,7 @@ package com.keynor.rpg.application.dto;
 
 import com.keynor.rpg.domain.model.BloodSystem;
 
-public record BloodSystemResponse(double oxygenCarryingCapacity) {
+public record BloodSystemResponse(int oxygenCarryingCapacity) {
 
     public static BloodSystemResponse from(BloodSystem bloodSystem) {
         return new BloodSystemResponse(bloodSystem.getOxygenCarryingCapacity());
