+++ src/main/java/com/keynor/rpg/application/dto/NervousSystemInput.java
@@ -2,7 +2,7 @@ package com.keynor.rpg.application.dto;
 
 import com.keynor.rpg.domain.model.NervousSystem;
 
-public record NervousSystemInput(double neuralDrive, double neuromuscularEfficiency) {
+public record NervousSystemInput(int neuralDrive, int neuromuscularEfficiency) {
 
     public NervousSystem toDomain() {
         return new NervousSystem(neuralDrive, neuromuscularEfficiency);
