+++ src/main/java/com/keynor/rpg/application/dto/PulmonarySystemInput.java
@@ -2,7 +2,7 @@ package com.keynor.rpg.application.dto;
 
 import com.keynor.rpg.domain.model.PulmonarySystem;
 
-public record PulmonarySystemInput(double pulmonaryCapacity) {
+public record PulmonarySystemInput(int pulmonaryCapacity) {
 
     public PulmonarySystem toDomain() {
         return new PulmonarySystem(pulmonaryCapacity);
