+++ src/main/java/com/keynor/rpg/application/dto/PulmonarySystemResponse.java
@@ -2,7 +2,7 @@ package com.keynor.rpg.application.dto;
 
 import com.keynor.rpg.domain.model.PulmonarySystem;
 
-public record PulmonarySystemResponse(double pulmonaryCapacity) {
+public record PulmonarySystemResponse(int pulmonaryCapacity) {
 
     public static PulmonarySystemResponse from(PulmonarySystem pulmonarySystem) {
         return new PulmonarySystemResponse(pulmonarySystem.getPulmonaryCapacity());
