+++ src/main/java/com/keynor/rpg/domain/model/BloodSystem.java
@@ -3,12 +3,15 @@ package com.keynor.rpg.domain.model;
 /**
  * Genetic — fixed once the character is created, unlike {@link CardiacSystem} and
  * {@link PulmonarySystem}. No setters: same immutability rule as {@link Genetics}.
+ *
+ * <p>Additive-standard discrete scale (rpg-11): 1-9, neutral 5 (formulas use the
+ * {@code value - 5} deviation).
  */
 public class BloodSystem {
 
-    private final double oxygenCarryingCapacity;
+    private final int oxygenCarryingCapacity;
 
-    public BloodSystem(double oxygenCarryingCapacity) {
+    public BloodSystem(int oxygenCarryingCapacity) {
         this.oxygenCarryingCapacity = oxygenCarryingCapacity;
     }
 
@@ -16,7 +19,7 @@ public class BloodSystem {
         return new BloodSystem(5);
     }
 
-    public double getOxygenCarryingCapacity() {
+    public int getOxygenCarryingCapacity() {
         return oxygenCarryingCapacity;
     }
 }
